fixed link colors
Fix allows users to change linkcolorhover, linkcolorfocus and linkcolorvisited from the variables.less file. In the original source code, the link color settings in typography.less did not match link colors settings in variables.less. This caused the link settings in the typography file to override the link settings in variables.less file.