sidebar image error
This code was already present in the gh-pages(as demo'd in https://mmistakes.github.io/minimal-mistakes/portfolio/foo-bar-website/), but after some looking around I found that the code that is being used by other users from the master repo has a bug while the gh-pages code for this file is correct. 
The bug is that when viewed on a mobile device or a smaller screen width for that matter, the sidebar doesn't stop the animation, so the image overlaps the text leaving the text not viewable.