disqus: Pass in canonical URL
Disqus will default to `window.location.href` if `this.page.url` is
unset and causes new comment threads to be created when the site is not
accesses from the canonical URL.

In my experience I have split threads for localhost Jekyll server use,
Google Translate and more.

To resolve this issue, pass along the canonical URL to Disqus so it can
create and serve the correct URLs.  The capture string was copied from
Discourse provider which appears to operate in the same manner.

More discussion: "JavaScript configuration variables"
https://help.disqus.com/customer/en/portal/articles/472098-javascript-configuration-variables