Changes the author-profile Schema so it is the "author" property.
This is so for the `single` layout, the `author` property is no longer a `Person` type
by itself, but it is now embedded inside the `CreativeWork` schema.

This means there is less repetition with the `Person` type that is in the
`_includes/seo.html` file.

To do this, it also moves the author_sidebar so it is inside the CreativeWork schema
in the `_layouts/single.html` file.