Uses `seo_author` Details if `site.social` is not filled in.
If `site.social.name` is not filled in for the Schema entry of type `Person` in
`seo.html`, the chosen `seo_author` field is used to fill it in instead.

In addition, more properties are filled in. However, only the social media
links for the social profiles supported by Google Search results
(see https://developers.google.com/search/docs/data-types/social-profile-links ) are
used.