Make excerpts optional in header overlay
Some posts and some header images don't lend themselves well to
displaying the excerpt in the overlay. Make this optional by
introducting a new boolean variable:

page.header.show_overlay_excerpt

Set it to implicitly default to true so existing users are unaffected by
testing explicitly for "!= false".

This can be enabled globally for a site by changing the default to false
in the local _config.yml, or per page by specifying the value in front
matter.

Signed-off-by: Darren Hart <darren@dvhart.com>