Allow specification of RSS icon in footer
All footer icons with the exception of the RSS icon can be customized in footer.html.   This update allows for the creation of a variable named site.atom_feed.icon in _config.yml in which the icon can be specified.   For example, the format would be something like:

atom_feed:
   icon: "fas fa-rss-square fa-2x"

If no value for this is specified, the current behavior would continue.